Invoke zoomed out when using the patterns tab, and create a hook
